pyproject.toml
[tool.poetry]
name = "diffusion-playground"
version = "0.1.0"
description = ""
authors = ["Jet <junjet.tai@sony.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.4.1"
numpy = "^2.1.1"
jj-wingman = "^0.22.2"
torchvision = "^0.19.1"
matplotlib = "^3.9.2"
tqdm = "^4.66.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
